Client/Server working. Although I can't seem to ad dmore functionality. I keep getting a cryptic error saying "unqualified enumeration constant name required." I'll need to ask Nick about that.
I've altered MCLPoseProvider and MCLParticleSet to start from a known pose, although I'm not exactly sure how well it is working. Whenever it moves, all the particles will have the same alterations. It could be a pointer problem, but I will need to investigate. 

Other than those two issues localization seems to be coming along. It can move and then scan every 30 degress around it. Once we get it working correctly, we will have to eperiment to compare number of scans and its accuracy to the length of time required for said scanning.